Move BadRequest to data level errors.py
Correct HTTPError handling by data.utils.SocrataApi.request
