+++ covid19_sfbayarea/data/errors.py
@@ -1,0 +1,10 @@
+import requests
+
+class BadRequest(requests.exceptions.HTTPError):
+    """
+    Represents a detailed error message from a web server.
+    """
+
+    def __str__(self):
+        message = super().__str__()
+        return f'{message} (status: {self.response.status_code})'
